whisper_server/backend/requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
sqlalchemy==2.0.41
pydantic==2.11.7
python-socketio==5.13.0
cryptography==45.0.4
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
PyJWT==2.10.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
Pillow==10.4.0
fastapi-mail==1.4.1
aiosmtplib==2.0.2
python-dotenv==1.0.0
aiosqlite